react-agent/01-github-pr-reviewer/requirements.txt
opperai>=0.28.0
aiohttp>=3.11.16
python-dotenv>=1.1.0
pydantic>=2.11.1